SSM-20140505: Added code to Copy Favourites to memory and re-insert them on a datbase update, so that they are not lost.  New Databases always have blank Favourites table.
Changed startup code so that calculation of branch distancesis performed
on the splash screen so that its available as soon as we get into the
app.

Created scripts for processing Branch information from the restaurant
websites.  This speeds up the import as its now done automatically.  Can
be problematic if Address formats are not uniform and if GPS location is
not readily available.  Relies on site to have checked data correctness.

Added all Gauteng stores for the five restaurants in the database.